(not building) Added Parser Generator
Learning about Compiler parsers. YACC / Lex tools are very abundant. So, it's just a matter of finding one for C# development after the C grammar is parsed. .lex and .y files have been added for ANSI C which are as old as 1985. There is one error using the YACC tools nuget which hopefully will get answered online in the next few days. After that the parser and analyzer code will generate automatically for the grammars in the ParserGenerator project. These will be used to start parsing C syntax in the UI setting!